feat: Replace all i64 to u32 and shift the map from (0, 10 000) to (100, 10 100)

+++ contracts/src/lib/shoot.cairo
@@ -3,7 +3,7 @@ use octoguns::models::bullet::{Bullet, BulletTrait};
 use dojo::world::{IWorldDispatcher, IWorldDispatcherTrait};
 
 fn shoot(world: IWorldDispatcher, bullet: Action, character: CharacterPosition) -> Bullet {
-    let angle = bullet.action_type; //i64
+    let angle = bullet.action_type; //u32
     let coords = character.coords; //Vec2
     let speed = 25;
     let id = world.uuid();
